update styles + create sceneUtils and refactor scenes
