+++ js/scenes.js
@@ -1,5 +1,6 @@
 import { sceneLayers } from './mapLayers.js'
 import customMap from './map.js'
+import * as utils from './sceneUtils.js'
 
 // get elements
 const scenes = document.querySelectorAll('.scene-section')
@@ -12,10 +13,10 @@ const mapThreeContainer = document.getElementById('map-3')
 const l = scenes.length
 const sceneObjs = []
 
-// get nav els minus hr
+// get navigable nav els
 const scrollNavBtns = Array.from(scrollNav).filter(el => el.nodeName != 'HR')
 
-// create map instances and store in indexed object
+// create and store map instances
 const maps = {
     stress: customMap(mapContainer),
     connectivity: customMap(mapTwoContainer),
@@ -32,12 +33,12 @@ for(let i = 0; i < l; i++) {
         triggerElement: scene,
         reverse: true
     })
-    .on('enter', e => {
-        toggleAnimation(e, scene)
-        toggleNavLink(i)
+    .on('enter', () => {
+        utils.toggleAnimation(scene)
+        utils.toggleNavLink(scrollNavBtns, i)
 
         if(mapId) {
-            const mapDiv = getMapCategory(scene)
+            const mapDiv = utils.getMapCategory(scene)
             const mapCategory = mapDiv.dataset.mapCategory
             toggleMapView(mapCategory, mapId)
         }
@@ -49,43 +50,13 @@ scrollNavBtns.forEach(btn => {
     let tooltip;
 
     btn.onmouseover = e => {
-        tooltip = makeNavTooltip(e)
+        tooltip = utils.makeNavTooltip(e)
         btn.insertAdjacentElement('afterend', tooltip)
     }
 
-    btn.onmouseout = () => removeNavTooltip(tooltip)
+    btn.onmouseout = () => utils.removeNavTooltip(tooltip)
 })
 
-// helper fncs to animate scene in/out, update side-nav link state and update map (when applicable)
-const toggleAnimation = (e, el) => {
-    const text = el.children[1] ? el.children[1] : el.children[0]
-    text.classList.add('fadein')
-}
-const toggleNavLink = i => {
-    // remove active class from all links
-    scrollNavBtns.forEach(btn => btn.classList.remove('scroll-story-nav-link-active'))
-
-    // add active class to current scene
-    scrollNavBtns[i].classList.add('scroll-story-nav-link-active')
-}
-const makeNavTooltip = e => {
-    let hrefToText = e.target.href.split('#')[1].split('-')
-    hrefToText.pop()
-    hrefToText = hrefToText.join(' ')
-    
-    const tooltipWrapper = document.createElement('span')
-    const tooltip = document.createElement('span')
-    
-    tooltipWrapper.classList.add('scroll-story-nav-tooltip-wrapper')
-    tooltip.classList.add('scroll-story-nav-tooltip')
-
-    tooltip.textContent = hrefToText
-
-    tooltipWrapper.appendChild(tooltip)
-    
-    return tooltipWrapper
-}
-const removeNavTooltip = tooltip => tooltip.remove()
 const toggleMapView = (mapCat, sceneId) => {
     let mapInstance;
     let mapSceneLayer;
@@ -110,12 +81,6 @@ const toggleMapView = (mapCat, sceneId) => {
 
     // apply map scene info from mapSceneLayer
     // zoom, center and filter are straightforward. layer could be complicated..
-    
-}
-const getMapCategory = el => {
-    const prev = el.previousElementSibling
-    if(prev.classList.contains('map')) return prev
-    else return getMapCategory(prev)
 }
 
 export default sceneObjs
